+++ src/commands/add/Add.ts
@@ -1,13 +1,20 @@
+import { Terminal } from 'vscode'
 import { Command } from '../../Command'
 import { showError, showMessage } from '../../utils/messages'
 import { getListInput } from '../../utils/selection'
 
 export class Add extends Command {
 
+  public static readonly terminalName = 'Angular Ng Add Package'
+  public static terminal?: Terminal
+
   private readonly popularPlugins: string[] = [
     '@angular/material',
     '@fortawesome/angular-fontawesome',
     '@ng-bootstrap/ng-bootstrap',
+    '@angular/fire',
+    '@azure/ng-deploy',
+    '@zeit/ng-deploy'
   ].sort()
 
   public async run() {
@@ -17,19 +24,11 @@ export class Add extends Command {
     }
 
     const command = `add ${addition}`
-    showMessage(`Attempting to add "${addition}". This may take a minute...`)
-    this.execCommand(command, ({ err }) => {
-      if (err) {
-        let message = 'The package could not be added.'
-        if (err.message.includes('does not support schematics')) {
-          message = 'The package does not support schematics and could not be installed.'
-        } else if (err.message.includes('404 Not Found')) {
-          message = 'The package does not exist on the registry.'
-        }
-        showError(message, err)
-      } else {
-        showMessage('The package was successfully added.')
-      }
-    })
+    showMessage(`Attempting to add "${addition}".`)
+    if (!Add.terminal) {
+      Add.terminal = this.execTerminal(Add.terminalName, command)
+    } else {
+      this.execTerminal(Add.terminal, command)
+    }
   }
 }