+++ src/commands/serve/Serve.ts
@@ -1,7 +1,7 @@
-import { Terminal } from 'vscode'
+import { env, Terminal, Uri } from 'vscode'
 import { Command } from '../../Command'
 import { showError, showMessage } from '../../utils/messages'
-import { getInput, getNoYes } from '../../utils/selection'
+import { getInput, getListInput, getNoYes } from '../../utils/selection'
 
 export class Serve extends Command {
 
@@ -18,19 +18,46 @@ export class Serve extends Command {
     Serve.host = await getInput('What host do you want to run the server with? (Default: localhost)', 'localhost')
     Serve.port = parseInt(await getInput('What port do you want to run the server on? (Default: 4200)', '4200'))
 
-    const openBrowser = await getNoYes('Do you want to open a browser window?')
+    const files = await this.getAngularFiles()
 
+    // Get a list of projects in the angular.json file
+    const projects = this.getProjects(files[0])
+    let project = ''
+    if (projects.length > 1) {
+      project = (await getListInput('Which project do you want to run?', projects)) || ''
+    } else if (projects.length === 1) {
+      project = projects[0]
+    }
+
+    // Get a list of configs for the project in the angular.json file
+    const projectConfigs = this.getProjectConfig(files[0], project)
+    let config = ''
+    if (projectConfigs.length > 1) {
+      config = (await getListInput('Which config do you want to use?', projectConfigs)) || ''
+    } else if (projectConfigs.length === 1) {
+      config = projectConfigs[0]
+    }
+
+    const ssl = await getNoYes('Do you need ssl?')
+
+    // Build the command
     const command = [
       `serve`,
+      project,
+      `--watch`,
       `--host=${Serve.host}`,
       `--port=${Serve.port}`,
-      `--watch`,
-      openBrowser ? '--open' : ''
+      ssl ? '--ssl' : '',
+      config.length > 0 ? `--configuration=${config}` : ''
     ]
 
-    Serve.server = this.execTerminal(Serve.terminalName, command.join(' '))
+    Serve.server = this.execTerminal(Serve.terminalName, command)
 
-    showMessage(`The server is starting on "http://${Serve.host}:${Serve.port}"`)
+    showMessage(`The server is starting on "http://${Serve.host}:${Serve.port}"`, 'Open Browser').then(button => {
+      if (button === 'Open Browser') {
+        env.openExternal(Uri.parse(`http://${Serve.host}:${Serve.port}`))
+      }
+    })
   }
 
   public stop() {