+++ CHANGELOG.md
@@ -1,2 +1,8 @@
 # 1.0.0
+- Add: `ng deploy`
+- Update: Added more items to `ng add` list
+- Update: Removed `Open in browser` question and added a button to the message to open in browser instead.
+- Fix: Integrated terminal freeze
+
+# 0.0.1
 - Initial release